web: Common: Structured files of PeopleSelector
