+++ web/ASC.Web.Common/src/components/people-selector/PeopleSelector.stories.js
